@@ -1,14 +1,14 @@
-import React, { useContext } from "react";
+import React, { useContext, useState, useEffect } from "react";
 
 import CartIcon from "../Cart/CartIcon";
 import Button from "../UI/Button";
 import CartContext from "../../store/cart-context";
 
 const HeaderCartButton = ({ onClickCart }) => {
+  const [animation, setAnimation] = useState(false);
   const { items } = useContext(CartContext);
 
-  console.log();
-
+  // Total item amount on header button
   const cartItemNumber = items.reduce((curr, item) => {
     return curr + item.amount;
   }, 0);
@@ -17,11 +17,23 @@ const HeaderCartButton = ({ onClickCart }) => {
     return onClickCart(true);
   };
 
+  useEffect(() => {
+    items.length === 0 ? setAnimation(false) : setAnimation(true);
+
+    const timer = setTimeout(() => {
+      setAnimation(false);
+    }, 700);
+
+    return () => {
+      clearTimeout(timer);
+    };
+  }, [items]);
+
   return (
     <Button
-      className={
-        "flex items-center justify-around gap-x-3 bg-[#4d1601] py-3 px-12 font-bold text-white hover:bg-[#2c0d00]"
-      }
+      className={`flex items-center justify-around gap-x-3 bg-[#4d1601] py-3 px-12 font-bold text-white hover:bg-[#2c0d00] ${
+        animation ? "animate-pulse" : ""
+      }`}
       type={"button"}
       onClick={cartButtonHandler}
     >
